feat(cardano): take the federation identity from the chain, and stop defaulting it (WI-069)
Y_federation and the recovery leaf's CSV delay build the Bitcoin treasury's Taproot address. Y_51
rotates each epoch with the DKG and arrives over the authenticated wire, but the federation leaf
does not rotate — every SPO rebuilds the tree locally, so a node holding a different value derives
a different scriptPubKey, signs a different BIP-341 sighash, and contributes a FROST share over a
message no other signer produced. Nothing errors: the address it derived is well-formed, it is
just not the one the BTC is in. Both were hand-copied from deployment notes into every operator's
config. They were already on chain the whole time — the treasury_info datum carries them at fields
#2/#3, heimdall writes them there at bootstrap, and treasury.ak enforces that copy for Update-Y —
so the fix reads what exists rather than publishing a second copy. The whole path hangs off the
one Config NFT: #12/#13 name the treasury_info policy and state NFT, its script hash gives the
address, and the snapshot heimdall already fetches to cross-check the identity trie carries the
datum. No new field, no extra query, no on-chain change.

The defaults are gone, which was the sharper half: y_fed_seed_hex defaulted to 0xFE x 32 and
federation_csv_blocks to 144, so an operator who set neither got no error, they got a confidently
derived wrong treasury. Both are Option, unset means "read the chain", and nowhere-configured is
fatal. federation_csv_blocks is now u16 end to end — it was a u32 that every consumer cast, so a
value above 65535 truncated into a different tree — and is validated where it enters. The identity
is resolved ONCE at startup and passed down, because the mover previously re-derived it from the
local seed and would have died on its first tick under the very config this change tells operators
to write. The seed keeps its one irreplaceable job, signing federation spends, so bootstrap-treasury,
federation-spend and the depositor demo still require it by name and share one parser; a blank
value now reads as unset. Startup check 9 reports the source, and the skip path lists every step,
which it did not before.

+++ src/epoch/machine.rs
@@ -688,12 +688,8 @@ async fn publish_keys_phase(
     match chain.query_treasury().await {
         Ok(treasury) => {
             let secp = Secp256k1::new();
-            let new_spend = treasury_spend_info(
-                &secp,
-                y_51,
-                treasury.y_fed,
-                treasury.federation_csv_blocks as u16,
-            );
+            let new_spend =
+                treasury_spend_info(&secp, y_51, treasury.y_fed, treasury.federation_csv_blocks);
             let new_spk = bitcoin::ScriptBuf::new_p2tr_tweaked(new_spend.output_key());
             crate::epoch_log!(
                 me,
@@ -1167,7 +1163,7 @@ async fn build_tm_phase(
         &secp,
         treasury.y_51,
         treasury.y_fed,
-        treasury.federation_csv_blocks as u16,
+        treasury.federation_csv_blocks,
     );
 
     // Treasury *change output*: send to the new roster's Taproot address,
@@ -1179,7 +1175,7 @@ async fn build_tm_phase(
         &secp,
         new_y_51,
         treasury.y_fed,
-        treasury.federation_csv_blocks as u16,
+        treasury.federation_csv_blocks,
     );
     let change_script = bitcoin::ScriptBuf::new_p2tr_tweaked(change_spend.output_key());
